Move methods from IClientMessenger to IServerStartupRemote
