+++ game-core/src/main/java/games/strategy/engine/framework/network/ui/ChangeGameToSaveGameClientAction.java
@@ -1,27 +1,56 @@
 package games.strategy.engine.framework.network.ui;
 
+import games.strategy.engine.framework.startup.mc.IServerStartupRemote;
 import games.strategy.engine.framework.startup.ui.panels.main.game.selector.GameFileSelector;
-import games.strategy.net.IClientMessenger;
 import java.awt.Frame;
 import java.awt.event.ActionEvent;
+import java.io.File;
+import java.io.IOException;
+import java.nio.file.Files;
+import java.util.logging.Level;
+import javax.annotation.Nonnull;
+import javax.annotation.Nullable;
 import javax.swing.AbstractAction;
+import lombok.extern.java.Log;
 
 /** An action for loading a save game across all network nodes from a client node. */
+@Log
 public class ChangeGameToSaveGameClientAction extends AbstractAction {
   private static final long serialVersionUID = -6986376382381381377L;
-  private final IClientMessenger clientMessenger;
+  private final IServerStartupRemote serverStartupRemote;
   private final Frame owner;
 
   public ChangeGameToSaveGameClientAction(
-      final IClientMessenger clientMessenger, final Frame owner) {
+      final IServerStartupRemote serverStartupRemote, final Frame owner) {
     super("Change To Gamesave (Load Game)");
-    this.clientMessenger = clientMessenger;
+    this.serverStartupRemote = serverStartupRemote;
     this.owner = owner;
   }
 
   @Override
   public void actionPerformed(final ActionEvent e) {
     GameFileSelector.selectGameFile(owner)
-        .ifPresent(file -> clientMessenger.changeToGameSave(file, file.getName()));
+        .ifPresent(file -> changeToGameSave(file, file.getName()));
+  }
+
+  public void changeToGameSave(@Nullable final File saveGame, final String fileName) {
+    final byte[] bytes = getBytesFromFile(saveGame);
+    if (bytes.length == 0) {
+      return;
+    }
+    serverStartupRemote.changeToGameSave(bytes, fileName);
+  }
+
+  @Nonnull
+  private static byte[] getBytesFromFile(@Nullable final File file) {
+    if (file == null || !file.exists()) {
+      return new byte[0];
+    }
+    try {
+      return Files.readAllBytes(file.toPath());
+    } catch (final IOException e) {
+      log.log(Level.SEVERE, "Failed to read file: " + file, e);
+      return new byte[0];
+    }
   }
 }
