+++ game-core/src/main/java/games/strategy/engine/framework/network/ui/SetMapClientAction.java
@@ -1,7 +1,6 @@
 package games.strategy.engine.framework.network.ui;
 
-import games.strategy.net.IClientMessenger;
-import games.strategy.net.INode;
+import games.strategy.engine.framework.startup.mc.IServerStartupRemote;
 import java.awt.Component;
 import java.awt.event.ActionEvent;
 import java.util.Collections;
@@ -17,29 +16,28 @@ public class SetMapClientAction extends AbstractAction {
 
   final List<String> availableGames;
   private final Component parent;
-  private final IClientMessenger clientMessenger;
+  private final IServerStartupRemote serverStartupRemote;
 
   public SetMapClientAction(
       final Component parent,
-      final IClientMessenger clientMessenger,
+      final IServerStartupRemote serverStartupRemote,
       final List<String> availableGames) {
     super("Change Game To");
     this.parent = JOptionPane.getFrameForComponent(parent);
-    this.clientMessenger = clientMessenger;
+    this.serverStartupRemote = serverStartupRemote;
     this.availableGames = availableGames;
     Collections.sort(this.availableGames);
   }
 
   @Override
   public void actionPerformed(final ActionEvent e) {
-    final INode serverNode = clientMessenger.getServerNode();
     final DefaultComboBoxModel<String> model = new DefaultComboBoxModel<>();
     final JComboBox<String> combo = new JComboBox<>(model);
     model.addElement("");
     for (final String game : availableGames) {
       model.addElement(game);
     }
-    if (serverNode == null || model.getSize() <= 1) {
+    if (model.getSize() <= 1) {
       JOptionPane.showMessageDialog(
           parent, "No available games", "No available games", JOptionPane.ERROR_MESSAGE);
       return;
@@ -54,6 +52,6 @@ public class SetMapClientAction extends AbstractAction {
     if (name == null || name.length() <= 1) {
       return;
     }
-    clientMessenger.changeServerGameTo(name);
+    serverStartupRemote.changeServerGameTo(name);
   }
 }
