+++ game-core/src/main/java/games/strategy/engine/framework/startup/mc/ClientModel.java
@@ -251,7 +251,6 @@ public class ClientModel implements IMessengerErrorListener {
     // save this, it will be cleared later
     gameDataOnStartup = gameSelectorModel.getGameData();
     final IServerStartupRemote serverStartup = getServerStartup();
-    messenger.setServerStartupRemote(serverStartup);
     final PlayerListing players = serverStartup.getPlayerListing();
     internalPlayerListingChanged(players);
     if (!serverStartup.isGameStarted(messenger.getLocalNode())) {
@@ -444,7 +443,7 @@ public class ClientModel implements IMessengerErrorListener {
   }
 
   public Action getHostBotSetMapClientAction(final Component parent) {
-    return new SetMapClientAction(parent, getClientMessenger(), getAvailableServerGames());
+    return new SetMapClientAction(parent, getServerStartup(), getAvailableServerGames());
   }
 
   public Action getHostBotChangeGameOptionsClientAction(final Component parent) {
@@ -452,12 +451,12 @@ public class ClientModel implements IMessengerErrorListener {
   }
 
   public Action getHostBotChangeGameToSaveGameClientAction(final Frame owner) {
-    return new ChangeGameToSaveGameClientAction(getClientMessenger(), owner);
+    return new ChangeGameToSaveGameClientAction(getServerStartup(), owner);
   }
 
   public Action getHostBotChangeToAutosaveClientAction(
       final Component parent, final HeadlessAutoSaveType autosaveType) {
-    return new ChangeToAutosaveClientAction(parent, getClientMessenger(), autosaveType);
+    return new ChangeToAutosaveClientAction(parent, getServerStartup(), autosaveType);
   }
 
   public Action getHostBotGetGameSaveClientAction(final Component parent) {
